simplified hardhat
