+++ packages/common/package.json
@@ -24,11 +24,17 @@
   ],
   "dependencies": {
     "@ethersproject/bignumber": "^5.0.5",
+    "@nomiclabs/hardhat-etherscan": "^2.1.0",
+    "@nomiclabs/hardhat-truffle5": "^2.0.0",
+    "@nomiclabs/hardhat-web3": "^2.0.0",
+    "@openzeppelin/contracts": "3.0.0",
     "@google-cloud/kms": "^0.3.0",
     "@google-cloud/storage": "^2.4.2",
     "@truffle/hdwallet-provider": "^1.0.25",
     "@umaprotocol/truffle-ledger-provider": "^1.0.3",
     "abi-decoder": "github:UMAprotocol/abi-decoder",
+    "hardhat": "^2.0.4",
+    "hardhat-gas-reporter": "^1.0.3",
     "bignumber.js": "^8.0.1",
     "dotenv": "^6.2.0",
     "eth-crypto": "^1.6.0",
