+++ packages/core/package.json
@@ -54,8 +54,8 @@
     "/config/*.json"
   ],
   "scripts": {
-    "test": "yarn mocha-test && yarn buidler-test",
-    "buidler-test": "node --max-old-space-size=1024 $(yarn bin buidler) test",
+    "test": "yarn mocha-test && yarn hardhat-test",
+    "hardhat-test": "node --max-old-space-size=1024 $(yarn bin hardhat) test",
     "mocha-test": "mocha mocha-test --recursive",
     "load-addresses": "yarn run apply-registry",
     "clean": "rm -rf build",
