+++ packages/disputer/hardhat.config.js
@@ -1,30 +1,17 @@
-require("@nomiclabs/hardhat-truffle5");
+const { getHardhatConfig } = require("@UMA/common");
+
 const path = require("path");
 const coreWkdir = path.dirname(require.resolve("@uma/core/package.json"));
-const liquidatorWkdir = path.dirname(require.resolve("@uma/liquidator/package.json"));
-module.exports = {
-  solidity: {
-    version: "0.6.12",
-    settings: {
-      optimizer: {
-        enabled: true,
-        runs: 199
-      }
-    }
-  },
-  networks: {
-    hardhat: {
-      gas: 11500000,
-      blockGasLimit: 11500000,
-      allowUnlimitedContractSize: false,
-      timeout: 1800000
-    }
-  },
+const packageWkdir = path.dirname(require.resolve("@uma/disputer/package.json"));
+
+const configOverride = {
   paths: {
     root: coreWkdir,
     sources: `${coreWkdir}/contracts`,
     artifacts: `${coreWkdir}/artifacts`,
     cache: `${coreWkdir}/cache`,
-    tests: `${liquidatorWkdir}/test`
+    tests: `${packageWkdir}/test`
   }
 };
+
+module.exports = getHardhatConfig(configOverride);
